+++ CHANGES.md
@@ -2,6 +2,7 @@
 * Added RSS feed icon next to social media icons
 * Added gallery support and combined it with fancybox
 * Added option to make galleries always flow well despite the window width
+* Changed multi-column layout style to single-column style and updated color schemes
 
 #### 1.3.5
 * Login page css fix